refactor: move add/get permission APIs into sdk
