+++ libs/sdk/src/api/index.ts
@@ -4,5 +4,6 @@ export * from "./base";
 export * from "./engine";
 export * from "./file";
 export * from "./insight";
+export * from "./permissions";
 export * from "./user";
 export * from "./websocket";
