+++ libs/shared/src/api/permissions.ts
@@ -103,74 +103,6 @@ export const getProjectUsers = async (
 };
 
 /**
- * Get users without access to a project
- * @param projectId - The project ID
- * @param admin - Whether to use admin endpoint
- * @param searchTerm - Optional search term to filter users by
- * @param limit - Optional limit for pagination
- * @param offset - Optional offset for pagination
- * @returns Array of users without project credentials
- */
-export const getProjectUsersNoCredentials = async (
-	projectId: string,
-	admin = false,
-	searchTerm?: string,
-	limit?: number,
-	offset?: number,
-): Promise<User[]> => {
-	let url = `${Env.MODULE}/api/auth/`;
-	if (admin) {
-		url += "admin/";
-	}
-
-	url += "project/getProjectUsersNoCredentials?";
-	url += `projectId=${projectId}`;
-	url += searchTerm ? `&searchTerm=${searchTerm}` : "";
-	url += limit !== undefined ? `&limit=${limit}` : "";
-	url += offset !== undefined ? `&offset=${offset}` : "";
-
-	const response = await get<User[]>(url).catch((error) => {
-		throw Error(error);
-	});
-
-	if (!response) {
-		throw Error("No Response to get non credentialed users");
-	}
-
-	return response.data;
-};
-
-/**
- * Add user permissions to a project
- * @param projectId - The project ID
- * @param users - Array of users with their permissions to add
- * @param admin - Whether to use admin endpoint
- * @returns Whether the operation was successful
- */
-export const addProjectUserPermissions = async (
-	projectId: string,
-	users: PostUser[],
-	admin = false,
-): Promise<boolean> => {
-	let url = `${Env.MODULE}/api/auth/`;
-	if (admin) {
-		url += "admin/";
-	}
-	url += "project/addProjectUserPermissions";
-
-	const response = await post<{ success: boolean }>(
-		url,
-		{
-			projectId,
-			userpermissions: users,
-		},
-		{},
-	);
-
-	return response.data.success;
-};
-
-/**
  * Edit user permissions for a project
  * @param projectId - The project ID
  * @param users - Array of users with their updated permissions
